+++ tests/test_oauth2.py
@@ -5,6 +5,7 @@ from urllib.parse import urlencode
 from urllib.parse import urlparse
 
 import pytest
+import responses
 
 from oic.oauth2 import Client
 from oic.oauth2 import Grant
@@ -18,11 +19,15 @@ from oic.oauth2.message import AccessTokenResponse
 from oic.oauth2.message import AuthorizationErrorResponse
 from oic.oauth2.message import AuthorizationRequest
 from oic.oauth2.message import AuthorizationResponse
+from oic.oauth2.message import CCAccessTokenRequest
 from oic.oauth2.message import DecodeError
 from oic.oauth2.message import ErrorResponse
+from oic.oauth2.message import ExtensionTokenRequest
 from oic.oauth2.message import FormatError
 from oic.oauth2.message import GrantExpired
+from oic.oauth2.message import MessageTuple
 from oic.oauth2.message import MissingRequiredAttribute
+from oic.oauth2.message import OauthMessageFactory
 from oic.oauth2.message import RefreshAccessTokenRequest
 from oic.utils import time_util
 from oic.utils.keyio import KeyBundle
@@ -68,10 +73,12 @@ class TestClient(object):
     def create_client(self):
         self.redirect_uri = "https://example.com/redirect"
         self.authorization_endpoint = "https://example.com/authz"
+        self.token_endpoint = "https://example.com/token"
 
         self.client = Client("1", config={"issuer": "https://example.com/as"})
         self.client.redirect_uris = [self.redirect_uri]
         self.client.authorization_endpoint = self.authorization_endpoint
+        self.client.token_endpoint = self.token_endpoint
 
     def test_construct_authz_req_no_optional_params(self):
         areq = self.client.construct_AuthorizationRequest(
@@ -181,6 +188,27 @@ class TestClient(object):
         assert atr["grant_type"] == "authorization_code"
         assert atr["code"] == "AbCdEf"
         assert atr["redirect_uri"] == self.redirect_uri
+
+    def test_construct_access_token_req_client_credentials(self):
+        # scope is default=""
+        request_args = {"grant_type": "client_credentials"}
+        atr = self.client.construct_AccessTokenRequest(
+            state="stat", request=CCAccessTokenRequest, request_args=request_args
+        )
+
+        assert atr["grant_type"] == "client_credentials"
+        assert atr["state"] == "stat"
+
+    def test_construct_access_token_req_extension_grant(self):
+        request_args = {
+            "grant_type": "urn:ietf:params:oauth:grant-type:saml2-bearer",
+            "assertion": "saml assertion",
+        }
+        atr = self.client.construct_AccessTokenRequest(
+            request=ExtensionTokenRequest, request_args=request_args
+        )
+        assert atr["grant_type"] == "urn:ietf:params:oauth:grant-type:saml2-bearer"
+        assert atr["assertion"] == "saml assertion"
 
     def test_construct_access_token_request_fail(self):
         with pytest.raises(GrantError):
@@ -520,6 +548,55 @@ class TestClient(object):
         with pytest.raises(MissingEndpoint):
             self.client._endpoint("token_endpoint")
             self.client._endpoint("foo_endpoint")
+
+    def test_do_access_token_request_client_credentials(self):
+        class CCMessageFactory(OauthMessageFactory):
+            """We are doing client credentials."""
+
+            token_endpoint = MessageTuple(CCAccessTokenRequest, AccessTokenResponse)
+
+        self.client.message_factory = CCMessageFactory
+        with responses.RequestsMock() as rsps:
+            rsps.add(
+                rsps.POST,
+                self.token_endpoint,
+                json={"access_token": "Token", "token_type": "bearer"},
+            )
+
+            resp = self.client.do_access_token_request()
+            assert rsps.calls[0].request.body == "grant_type=client_credentials"
+
+        assert isinstance(resp, AccessTokenResponse)
+        assert resp["access_token"] == "Token"
+
+    def test_do_access_token_request_extension_grant(self):
+        class ExtensionMessageFactory(OauthMessageFactory):
+            """We are doing Extension grant."""
+
+            token_endpoint = MessageTuple(ExtensionTokenRequest, AccessTokenResponse)
+
+        self.client.message_factory = ExtensionMessageFactory
+        request_args = {
+            "assertion": "saml assertion",
+            "grant_type": "urn:ietf:params:oauth:grant-type:saml2-bearer",
+        }
+        with responses.RequestsMock() as rsps:
+            rsps.add(
+                rsps.POST,
+                self.token_endpoint,
+                json={"access_token": "Token", "token_type": "bearer"},
+            )
+
+            resp = self.client.do_access_token_request(request_args=request_args)
+            request = parse_qs(rsps.calls[0].request.body)
+            assert request["assertion"][0] == "saml assertion"
+            assert (
+                request["grant_type"][0]
+                == "urn:ietf:params:oauth:grant-type:saml2-bearer"
+            )
+
+        assert isinstance(resp, AccessTokenResponse)
+        assert resp["access_token"] == "Token"
 
 
 class TestServer(object):